+++ lib/src/backend.rs
@@ -21,8 +21,10 @@ use std::slice;
 use std::time::SystemTime;
 
 use async_trait::async_trait;
-use chrono::TimeZone as _;
 use futures::stream::BoxStream;
+use jiff::Zoned;
+use jiff::tz::Offset;
+use jiff::tz::TimeZone;
 use thiserror::Error;
 use tokio::io::AsyncRead;
 
@@ -78,7 +80,13 @@ impl CopyId {
 
 #[derive(Debug, Error)]
 #[error("Out-of-range date")]
-pub struct TimestampOutOfRange;
+pub struct TimestampOutOfRange(pub jiff::Error);
+
+impl From<TimestampOutOfRange> for jiff::Error {
+    fn from(value: TimestampOutOfRange) -> Self {
+        value.0
+    }
+}
 
 #[derive(ContentHash, Debug, PartialEq, Eq, Clone, Copy, PartialOrd, Ord)]
 pub struct MillisSinceEpoch(pub i64);
@@ -92,36 +100,22 @@ pub struct Timestamp {
 
 impl Timestamp {
     pub fn now() -> Self {
-        Self::from_datetime(chrono::offset::Local::now())
+        Self::from_zoned(Zoned::now())
     }
 
-    pub fn from_datetime<Tz: chrono::TimeZone<Offset = chrono::offset::FixedOffset>>(
-        datetime: chrono::DateTime<Tz>,
-    ) -> Self {
+    pub fn from_zoned(zoned: Zoned) -> Self {
         Self {
-            timestamp: MillisSinceEpoch(datetime.timestamp_millis()),
-            tz_offset: datetime.offset().local_minus_utc() / 60,
+            timestamp: MillisSinceEpoch(zoned.timestamp().as_millisecond()),
+            tz_offset: zoned.time_zone().to_offset(zoned.timestamp()).seconds() / 60,
         }
     }
 
-    pub fn to_datetime(
-        &self,
-    ) -> Result<chrono::DateTime<chrono::FixedOffset>, TimestampOutOfRange> {
-        let utc = match chrono::Utc.timestamp_opt(
-            self.timestamp.0.div_euclid(1000),
-            (self.timestamp.0.rem_euclid(1000)) as u32 * 1000000,
-        ) {
-            chrono::LocalResult::None => {
-                return Err(TimestampOutOfRange);
-            }
-            chrono::LocalResult::Single(x) => x,
-            chrono::LocalResult::Ambiguous(y, _z) => y,
-        };
-
-        Ok(utc.with_timezone(
-            &chrono::FixedOffset::east_opt(self.tz_offset * 60)
-                .unwrap_or_else(|| chrono::FixedOffset::east_opt(0).unwrap()),
-        ))
+    pub fn to_zoned(&self) -> Result<Zoned, TimestampOutOfRange> {
+        Ok(jiff::Timestamp::from_millisecond(self.timestamp.0)
+            .map_err(TimestampOutOfRange)?
+            .to_zoned(TimeZone::fixed(
+                Offset::from_seconds(self.tz_offset * 60).map_err(TimestampOutOfRange)?,
+            )))
     }
 }
 
@@ -131,8 +125,8 @@ impl serde::Serialize for Timestamp {
         S: serde::Serializer,
     {
         // TODO: test is_human_readable() to use raw format?
-        let t = self.to_datetime().map_err(serde::ser::Error::custom)?;
-        t.serialize(serializer)
+        let t = self.to_zoned().map_err(serde::ser::Error::custom)?;
+        t.timestamp().to_string().serialize(serializer)
     }
 }